Issue #202: Reduce the file length of InputMutationTests4
# Conflicts:
#	config/checkstyle-resources-suppressions.xml
#	src/test/java/org/checkstyle/autofix/recipe/UseEnhancedSwitchTest.java
