+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests26/DiffMutationTests26.diff
@@ -1,0 +1,92 @@
+--- src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests26/InputMutationTests26.java
++++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests26/OutputMutationTests26.java
+@@ -8,66 +8,49 @@
+ 
+ package org.checkstyle.autofix.recipe.useenhancedswitch.mutationtests26;
+ 
+-public class InputMutationTests26 {
++public class OutputMutationTests26 {
+ 
+     void assignmentArrowBlock(int x) {
+-        int val;
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (x) {
+-            case 1: {
+-                val = 10;
+-                break;
+-            }
+-            case 2: {
++        int val = switch (x) {
++            case 1 -> 10;
++            case 2 -> {
+                 System.out.println("two");
+-                val = 20;
+-                break;
++                yield 20;
+             }
+-            default: {
+-                val = 0;
+-                break;
+-            }
+-        }
++            default -> 0;
++        };
+         System.out.println(val);
+     }
+ 
+     int blockReturnExpr(int x) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (x) {
+-            case 1:
++        return switch (x) {
++            case 1 -> {
+                 System.out.println("case1");
+-                return 1;
+-            case 2:
+-                return 2;
+-            default:
++                yield 1;
++            }
++            case 2 -> 2;
++            default -> {
+                 System.out.println("default");
+-                return 0;
+-        }
++                yield 0;
++            }
++        };
+     }
+ 
+     void assignmentWithThrow(int x) {
+-        String val;
+-        // violation below, 'Switch can be replaced with enhanced switch'
+-        switch (x) {
+-            case 1:
+-                val = "one";
+-                break;
+-            default:
+-                throw new RuntimeException("error");
+-        }
++        String val = switch (x) {
++            case 1 -> "one";
++            default -> throw new RuntimeException("error");
++        };
+         System.out.println(val);
+     }
+ 
+     void multiStatement(int x) {
+-        // violation below, 'Switch can be replaced with enhanced switch'
+         switch (x) {
+-            case 1:
++            case 1 -> {
+                 System.out.println("a");
+                 System.out.println("b");
+-                break;
+-            default:
+-                System.out.println("default");
+-                break;
++            }
++            default -> System.out.println("default");
+         }
+     }
+ }
